Moving handle_hooks to try and make #708 easier to review (#724)
